+++ OelApp/Model/Session.cs
@@ -4,7 +4,7 @@ namespace OelApp
 {
     public class Session
     {
-        public int NumberOfUnits { get; set; }
+        public double NumberOfUnits { get; set; }
 
         public DateTime StartTime { get; set; }
         public Person Person { get; set; }
